+++ prisma/migrations/20220211065535_remove_unique_from_address_id_in_company_rtable/migration.sql
@@ -1,0 +1,6 @@
+-- DropIndex
+DROP INDEX "Company_addressId_key";
+
+-- AlterTable
+ALTER TABLE "Lead" ALTER COLUMN "lastCommunication" DROP NOT NULL,
+ALTER COLUMN "lastCommunication" SET DATA TYPE DATE;