+++ src/address/dto/update-address.dto.ts
@@ -1,0 +1,4 @@
+import { PartialType } from '@nestjs/mapped-types';
+import { CreateaddressDto } from './create-address.dto';
+
+export class UpdateaddressDto extends PartialType(CreateaddressDto) {}