+++ src/company/dto/create-company.dto.ts
@@ -1,0 +1,22 @@
+import {
+  IsDefined,
+  IsEmpty,
+  IsNumber,
+  IsOptional,
+  IsString,
+} from 'class-validator';
+
+export class CreateCompanyDto {
+  @IsDefined()
+  @IsString()
+  name: string;
+
+  @IsOptional()
+  @IsNumber()
+  addressId: number;
+
+  @IsEmpty()
+  updatedAt: Date;
+  @IsEmpty()
+  createdAt: Date;
+}