+++ src/lead/lead.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { LeadService } from './lead.service';
 import { LeadController } from './lead.controller';
-import { PrismaService } from '../prisma.services';
 
 @Module({
   controllers: [LeadController],
-  providers: [LeadService, PrismaService],
+  providers: [LeadService],
 })
 export class LeadModule {}