+++ src/lead/lead.service.ts
@@ -1,14 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { Lead, Prisma } from '@prisma/client';
-import { PrismaService } from '../prisma.services';
+import { PrismaService } from 'src/prisma/prisma.service';
+import { UpdateCreateDto } from './dto/update-lead.dto';
 
 @Injectable()
 export class LeadService {
   constructor(private prisma: PrismaService) {}
 
   create(leadData: Prisma.LeadCreateInput) {
-    this.prisma.lead.create({ data: leadData });
-    return;
+    if (leadData.lastCommunication) {
+      leadData.lastCommunication = new Date(leadData.lastCommunication);
+    }
+
+    return this.prisma.lead.create({ data: leadData });
   }
 
   findAll() {
@@ -18,10 +22,18 @@ export class LeadService {
   async findOneById(id: Prisma.LeadWhereUniqueInput): Promise<Lead | null> {
     return this.prisma.lead.findUnique({
       where: id,
+      include: {
+        company: {
+          include: {
+            address: true,
+          },
+        },
+      },
     });
   }
 
-  update(id: Prisma.LeadWhereUniqueInput, data: Prisma.LeadUpdateInput) {
+  update(id: Prisma.LeadWhereUniqueInput, data: UpdateCreateDto) {
+    data.updatedAt = new Date();
     return this.prisma.lead.update({ where: id, data });
   }