+++ src/prisma/prisma.service.ts
@@ -1,8 +1,11 @@
-import { INestApplication, Injectable, OnModuleInit } from '@nestjs/common';
+import { INestApplication, Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 
 @Injectable()
-export class PrismaService extends PrismaClient implements OnModuleInit {
+export class PrismaService extends PrismaClient {
+  constructor() {
+    super({ log: ['info'] });
+  }
   async onModuleInit() {
     await this.$connect();
   }